requirements.txt
Django==1.6
PIL==1.1.7
South==0.8.4
argparse==1.2.1
beautifulsoup4==4.3.2
boto==2.19.0
dj-database-url==0.2.2
django-storages==1.1.8
django-tinymce==1.5.1
psycopg2==2.5.1
wsgiref==0.1.2
